this is the first code of tic_tac_toe
Hey,

The tic_tac_toe is a simple game, here it is for the 2 players. Actually the game rules are if same player have complete row entry or column entry or any diagonal three elements are alike then he is the winner. Below I am explaining the game

Player 1 value is "X" 
Player 2 value is "O"
and the entries will be like this 1 | 2 | 3
                                                  4 | 5 | 6
                                                  7 | 8 | 9
here 1,2,3,4,5,6,7,8,9 are the positions have to choose, each player have to choose positions one after another.

example:
Player 1 : 2  
                 1 | X | 3
                 4 | 5 | 6
                 7 | 8 | 9
Player 2 : 4
                 1 | X | 3
                O | 5 | 6
                 7 | 8 | 9
like this